refactor: migrate GenderField folder to typescript
(cherry picked from commit a2d34b3ac986fa6b7f4dce7b8933cc2048219486)
