--- frontend/src/components/form/GenderField/GenderField.stories.js
@@ -1,12 +1,0 @@
-import React from "react";
-import GenderField from "./GenderField";
-
-export default {
-  title: "form/GenderField",
-  component: GenderField,
-};
-
-const Template = (args) => <GenderField {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
