+++ frontend/src/components/form/GenderField/GenderField.tsx
@@ -1,11 +1,16 @@
 import classNames from "classnames";
-import PropTypes from "prop-types";
-import React from "react";
 import Label from "../../@common/Label/Label";
 import Radio from "../../@common/Radio/Radio";
 import styles from "./GenderField.module.css";
 
-const GenderField = ({ required, className, value, onChange }) => {
+type GenderFieldProps = {
+  className?: string;
+  required?: boolean;
+  value: string;
+  onChange: React.ChangeEventHandler<HTMLInputElement>;
+};
+
+const GenderField = ({ className, required = false, value, onChange }: GenderFieldProps) => {
   return (
     <div className={classNames(styles["gender-field"], className)}>
       <Label className={styles.label} required={required}>
@@ -31,17 +36,6 @@ const GenderField = ({ required, className, value, onChange }) => {
       </div>
     </div>
   );
-};
-
-GenderField.propTypes = {
-  value: PropTypes.string.isRequired,
-  onChange: PropTypes.func.isRequired,
-  required: PropTypes.bool,
-  className: PropTypes.string,
-};
-
-GenderField.defaultProps = {
-  required: false,
 };
 
 export default GenderField;
